ParsedJson::Iterator is now a template class.
